[syslog] Use universal printf type on 32/ 64 bit builds
Symptom: Always build failed.
Root cause: Type diff.
Solution: Use %zu for size_t, PRIu64 for uint64_t.
Note: NA

+++ src/core.cpp
@@ -114,11 +114,10 @@ inline int checkIsBufferOverflow(const std::vector<CviImg> &input,
     uint32_t total_addr = is_1d ? input[k].GetImgSize() : input[k].m_tg.stride.c;
 #endif
     if (jumped_value != total_addr) {
-      LOGE(
-          "[%u] Error! Input %u jumped value %lu not align to image size %u, start addr "
-          "%lu\n",
-          (uint32_t)b, (uint32_t)k, (long unsigned int)jumped_value, total_addr,
-          (long unsigned int)bm_start_addr);
+      LOGE("[%zu] Error! Input %zu jumped value %" PRIu64
+           " not align to image size %u, start addr "
+           "%" PRIu64 "\n",
+           b, k, jumped_value, total_addr, bm_start_addr);
       ret = CVI_FAILURE;
     }
   }
@@ -139,11 +138,10 @@ inline int checkIsBufferOverflow(const std::vector<CviImg> &input,
     uint32_t total_addr = (is_1d ? output[k].GetImgSize() : output[k].m_tg.stride.c) + pad_offset;
 #endif
     if (jumped_value != total_addr) {
-      LOGE(
-          "[%u] Error! Output %u jumped value %lu not align to image size %u, start addr "
-          "%lu\n",
-          (uint32_t)b, (uint32_t)k, (long unsigned int)jumped_value, total_addr,
-          (long unsigned int)bm_des_addr);
+      LOGE("[%zu] Error! Output %zu jumped value %" PRIu64
+           " not align to image size %u, start addr "
+           "%" PRIu64 "\n",
+           b, k, jumped_value, total_addr, bm_des_addr);
       ret = CVI_FAILURE;
     }
   }
@@ -1108,14 +1106,14 @@ int IveCore::runSingleSizeExtKernel(CVI_RT_HANDLE rt_handle, cvk_context_t *cvk_
           tg_in.fmt = bm_src_info.fns_vec[k].getFmt();
 
           // clang-format off
-          LOGD("[%u] In\n"
-               " tg start address %lu\n"
+          LOGD("[%zu] In\n"
+               " tg start address %" PRIu64 "\n"
                " tg shape %d %d %d %d\n"
                " tg stride %d %d %d\n"
                " tl shape %d %d %d %d\n"
                " tl stride %u %u %u %u\n",
-               (uint32_t)k,
-               (long unsigned int)tg_in.start_address,
+               k,
+               tg_in.start_address,
                tg_in.shape.n, tg_in.shape.c, tg_in.shape.h, tg_in.shape.w,
                tg_in.stride.n, tg_in.stride.c, tg_in.stride.h,
                tl_in[k]->shape.n, tl_in[k]->shape.c, tl_in[k]->shape.h, tl_in[k]->shape.w,
@@ -1156,14 +1154,14 @@ int IveCore::runSingleSizeExtKernel(CVI_RT_HANDLE rt_handle, cvk_context_t *cvk_
           out_shape.shape.w = tg_out.shape.w;
 
           // clang-format off
-          LOGD("[%u] Out\n"
-               " tg start address %lu\n"
+          LOGD("[%zu] Out\n"
+               " tg start address %" PRIu64 "\n"
                " tg shape %d %d %d %d\n"
                " tg stride %d %d %d\n"
                " tl shape %d %d %d %d\n"
                " tl stride %u %u %u %u\n",
-               (uint32_t)k,
-               (long unsigned int)tg_out.start_address,
+               k,
+               tg_out.start_address,
                tg_out.shape.n, tg_out.shape.c, tg_out.shape.h, tg_out.shape.w,
                tg_out.stride.n, tg_out.stride.c, tg_out.stride.h,
                tl_out[k]->shape.n, tl_out[k]->shape.c, tl_out[k]->shape.h, tl_out[k]->shape.w,
@@ -1310,7 +1308,7 @@ int IveCore::runNoKernel(CVI_RT_HANDLE rt_handle, cvk_context_t *cvk_ctx,
     shape.w = w_val;
   }
   LOGD("Total size %u\n", total_size);
-  LOGD("turn %u left %u\n", (uint32_t)loop_turn, (uint32_t)left_pixels);
+  LOGD("turn %zu left %zu\n", loop_turn, left_pixels);
   LOGD("%u %u %u %u\n", shape.n, shape.c, shape.h, shape.w);
 
   std::vector<cvk_tg_shape_t> s_in_vec, s_out_vec;
@@ -1366,7 +1364,7 @@ int IveCore::runNoKernel(CVI_RT_HANDLE rt_handle, cvk_context_t *cvk_ctx,
         p_copy_in.src = &tg_in;
         p_copy_in.dst = tl_in_info.lmem_vec[k + pp_skip];
         cvk_ctx->ops->tdma_g2l_bf16_tensor_copy(cvk_ctx, &p_copy_in);
-        LOGD("tg_in physical addr %lu\n", (long unsigned int)tg_in.start_address);
+        LOGD("tg_in physical addr %" PRIu64 "\n", tg_in.start_address);
         // Change src head addr
         bm_src_info.addr_vec[k] += 1 * jump_src * bm_src_info.fns_vec[k].getSize();
       }
@@ -1392,7 +1390,7 @@ int IveCore::runNoKernel(CVI_RT_HANDLE rt_handle, cvk_context_t *cvk_ctx,
         p_copy_out.src = tl_out_info.lmem_vec[k + pp_skip];
         p_copy_out.dst = &tg_out;
         cvk_ctx->ops->tdma_l2g_bf16_tensor_copy(cvk_ctx, &p_copy_out);
-        LOGD("tg_out physical addr %lu\n", (long unsigned int)tg_out.start_address);
+        LOGD("tg_out physical addr %" PRIu64 "\n", tg_out.start_address);
         // Change dest head addr
         bm_dest_info.addr_vec[k] += 1 * jump_dst * bm_dest_info.fns_vec[k].getSize();
       }
@@ -1469,7 +1467,7 @@ int IveCore::runNoKernel(CVI_RT_HANDLE rt_handle, cvk_context_t *cvk_ctx,
       p_copy_in.src = &tg_in;
       p_copy_in.dst = tl_in_info.lmem_vec[k];
       cvk_ctx->ops->tdma_g2l_bf16_tensor_copy(cvk_ctx, &p_copy_in);
-      LOGD("tg_in physical addr %lu\n", (long unsigned int)tg_in.start_address);
+      LOGD("tg_in physical addr %" PRIu64 "\n", tg_in.start_address);
       // Change src head addr
       bm_src_info.addr_vec[k] += 1 * jump_src * bm_src_info.fns_vec[k].getSize();
     }
@@ -1490,7 +1488,7 @@ int IveCore::runNoKernel(CVI_RT_HANDLE rt_handle, cvk_context_t *cvk_ctx,
       p_copy_out.src = tl_out_info.lmem_vec[k];
       p_copy_out.dst = &tg_out;
       cvk_ctx->ops->tdma_l2g_bf16_tensor_copy(cvk_ctx, &p_copy_out);
-      LOGD("tg_out physical addr %lu\n", (long unsigned int)tg_out.start_address);
+      LOGD("tg_out physical addr %" PRIu64 "\n", tg_out.start_address);
       // Change dest head addr
       bm_dest_info.addr_vec[k] += 1 * jump_dst * bm_dest_info.fns_vec[k].getSize();
     }
